refactor: encapsulated image loading
