+++ src/lib/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './errors';
+export * from './post-hero-image';
